remove the `config.assets_file` options and handling
This was a relic of previous implementations back when a `dassets`
CLI was part of the tool design.  This is no longer the case and
this config is not needed.
