+++ test/helper.rb
@@ -11,6 +11,23 @@ require 'pathname'
 TEST_SUPPORT_PATH = Pathname.new(File.expand_path('../support', __FILE__))
 
 ENV['DASSETS_TEST_MODE']   = 'yes'
-ENV['DASSETS_ASSETS_FILE'] = 'test/support/config/assets'
+
 require 'dassets'
+
+@dumb_engine = Class.new(Dassets::Engine) do
+  def ext(in_ext); ''; end
+  def compile(input); "#{input}\nDUMB"; end
+end
+@useless_engine = Class.new(Dassets::Engine) do
+  def ext(in_ext); 'no-use'; end
+  def compile(input); "#{input}\nUSELESS"; end
+end
+
+Dassets.configure do |c|
+  c.source TEST_SUPPORT_PATH.join("app/assets") do |s|
+    s.engine 'dumb', @dumb_engine
+    s.engine 'useless', @useless_engine
+  end
+end
+
 Dassets.init
